day4/folder1/requirements.txt
streamlit
openai-agents
tavily-python
openai
pydantic
python-dotenv
